requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp==3.7.4.post0
altair==4.1.0
appdirs==1.4.4
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828452693/work
args==0.1.0
astor==0.8.1
astroid==2.5.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
audiomentations==0.17.0
audioread==2.1.9
autopep8 @ file:///tmp/build/80754af9/autopep8_1613246362269/work
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
boto3==1.18.16
botocore==1.21.16
Box2D==2.3.10
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
certifi==2020.12.5
cffi==1.14.0
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
click==7.1.2
clint==0.5.1
cloudpickle==1.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1607635305226/work
cssselect==1.1.0
cupy-cuda101==8.5.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.4
discord.py==1.7.3
entrypoints==0.3
fastrlock==0.5
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
flake8 @ file:///tmp/build/80754af9/flake8_1601911421857/work
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
gast==0.3.3
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.18
google-auth @ file:///tmp/build/80754af9/google-auth_1600960338579/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424467487/work
gym==0.18.0
h5py @ file:///tmp/build/80754af9/h5py_1593454119955/work
huggingface-hub==0.0.12
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
inflect==5.3.0
ipdb==0.13.4
ipykernel==5.4.2
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==2.0.1
jedi @ file:///tmp/build/80754af9/jedi_1606932558467/work
jieba3k==0.35.1
Jinja2==3.0.1
jmespath==0.10.0
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
json5==0.9.5
jsonpatch==1.32
jsonpointer==2.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148959479/work
jupyterlab==2.2.6
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
jupyterlab-widgets==1.0.0
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Preprocessing==1.1.0
keyboard==0.13.5
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
lameenc==1.3.1
lazy-object-proxy==1.6.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.6.3
Markdown @ file:///tmp/build/80754af9/markdown_1605111062842/work
MarkupSafe==2.0.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603376012865/work
mccabe==0.6.1
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
multiprocess==0.70.12.2
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914821128/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
newspaper3k==0.2.8
nltk==3.5
notebook @ file:///tmp/build/80754af9/notebook_1608835548841/work
numba==0.53.1
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.1.0
packaging==21.0
pandas==1.2.1
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==7.2.0
pooch==1.4.0
processing-py==0.3.7
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit==3.0.16
protobuf==3.13.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.6.2
pydot==1.4.1
pydub==0.25.1
pyflakes==2.2.0
pyglet==1.5.0
Pygments @ file:///tmp/build/80754af9/pygments_1607368905949/work
PyJWT==1.7.1
pylint==2.7.4
pyloudnorm==0.1.0
pynndescent==0.5.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141707582/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.1
python-dotenv==0.18.0
PyTrie==0.4.0
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-file==1.5.1
requests-oauthlib==1.3.0
resampy==0.2.2
rope @ file:///tmp/build/80754af9/rope_1602264064449/work
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376882706/work
scipy==1.4.1
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sgmllib3k==1.0.0
six @ file:///tmp/build/80754af9/six_1605205313296/work
smart-open==4.2.0
smmap==4.0.0
sortedcontainers==2.3.0
sounddevice==0.4.2
SoundFile==0.10.3.post1
soupsieve==2.2.1
sshkeys==0.5
streamlit==0.86.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.9.0
tensorflowjs==2.8.3
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tinysegmenter==0.3
tldextract==3.1.0
tmuxpair==1.0.0
tokenizers==0.10.3
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.11.1
torch==1.8.1
torchaudio==0.8.1
torchfile==0.1.0
torchvision==0.9.1
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
transformers==4.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
umap-learn==0.5.1
Unidecode==1.2.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
validators==0.18.2
visdom==0.1.8.9
watchdog==2.1.3
wavio==0.0.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
yarl==1.6.3
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
